04_assistant_v2_tools/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
langsmith>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pytest>=8.0.0
typer>=0.12.0
rich>=13.7.0
